feat: modify intro behavior
