+++ website/app/layout.tsx
@@ -20,7 +20,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en" className={inter.className}>
+    <html lang="en" className={inter.className} style={{scrollBehavior:'smooth'}}>
       <body>
         <Theme {...themeProps}>
           <main>{children}</main>
